+++ CHANGELOG.markdown
@@ -1,22 +1,39 @@
+/1.3/
+
+ * Removed old benchmark/test framework.
+ * Make numSparks support mandatory.
+ * Moved to Control.Concurrent from Data
+ * Added MonadSpec, so we can make instances for Codensity, etc. in other packages
+ * Removed the ContT r STM combinators
+ * Cleaned out old issues
+ * Made compatible with the removal of the Eq/Show superclasses of Num in GHC 7.3+
+
 /1.2.0.2/
+
  * Fixed name collision with the new Distribution.Simple.testHook in Setup.lhs
 
 /1.2.0.1/
+
  * Weakened dependencies
 
 /1.2.0.0/:
- * Reorganized the module hierarchy into Data.Speculation.*
+
+ * Reorganized the module hierarchy into Data.Speculation
 
 /1.1.0.0/:
+
  * Added support for numSparks
 
 /1.0.0.0/:
+
  * Released
 
 /0.9.0/:
+
  * Removed interim boxing in the unsafeIsEvaluated and unsafeGetTagBits check
 
 /0.8.1/:
+
  * Added Data.List.Foldable
  * Added Data.Traversable.Foldable
  * Fixed an off-by-one error in the arguments to the speculative fold estimators
@@ -26,5 +43,5 @@
  * changed tests and benchmarks to not build by default to work around corruption in the hackage db
 
 /0.8.0.1/:
- 
+
  * test suite now forces build